django_ecommerce/mobile_service/requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
djongo==1.3.6
pymongo==3.12.3
requests>=2.25.0 